refactor: update amount types from int to double across various components and enhance number formatting

+++ lib/core/widgets/charts/app_chart.bar.dart
@@ -36,8 +36,8 @@ class AppChartBarState extends State<AppChartBar> {
             final monthly = frequency == 'Monthly';
             final yearly = frequency == 'Yearly';
 
-            var dailyAmounts = <int>[];
-            var monthlyAmounts = <int>[];
+            var dailyAmounts = <double>[];
+            var monthlyAmounts = <double>[];
 
             if (dailyTransactions.isNotEmpty) {
               dailyAmounts = dailyTransactions.values.toList();
